backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
click==8.3.2
deepgram-sdk==6.1.1
Deprecated==1.3.1
distro==1.9.0
fastapi==0.136.0
frozenlist==1.8.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.12
limits==5.8.0
multidict==6.7.1
packaging==26.1
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dotenv==1.2.2
python-magic==0.4.27
python-multipart==0.0.26
PyYAML==6.0.3
slowapi==0.1.9
sniffio==1.3.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.45.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
yt-dlp==2026.3.17
